+++ contracts/Token/Token.sol
@@ -9,12 +9,12 @@ import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
  * @dev IERC20 used for test purposes
  */
 contract Token is ERC20 {
-    uint8 private immutable _decimals;
+    uint256 private immutable _decimals;
 
     constructor(
         string memory name_,
         string memory symbol_,
-        uint8 decimals_,
+        uint256 decimals_,
         address receiver_,
         uint totalSupply_
     ) ERC20(name_, symbol_) {
@@ -22,7 +22,7 @@ contract Token is ERC20 {
         _mint(receiver_, totalSupply_ * (10 ** decimals_));
     }
 
-    function decimals() public view virtual override returns (uint8) {
+    function decimals() public view virtual override returns (uint256) {
         return _decimals;
     }
 }
